Update categories fetching in search-input component
